+++ src/main/java/nl/novi/backend/spring/api/kerkapp/Repository/BibleRepository.java
@@ -1,9 +1,10 @@
 package nl.novi.backend.spring.api.kerkapp.Repository;
 
-import nl.novi.backend.spring.api.kerkapp.Entittiy.Bible;
+import nl.novi.backend.spring.api.kerkapp.Entitiy.Bible;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public interface BibleRepository extends JpaRepository<Bible, Long> {
+
 }
